Move tests of math_utils to separate test file.
